requirements.txt
langchain-community
bs4
pymongo[srv]
python-dotenv
uagents
langchain-groq
uagents-ai-engine